Autenticação com github e criação de JWT, criação de middleware e proteção de rotas privadas

+++ src/routes/memories.ts
@@ -2,11 +2,16 @@ import { FastifyInstance } from 'fastify'
 import { prisma } from '../lib/prisma'
 import { z } from 'zod'
 
-const userId = '02f634ee-de94-4af2-8e82-e3909428d73a';
-
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify();
+  })
+
+  app.get('/memories', async (request) => {
+    const userId = request.user.sub;
+
     const memories = await prisma.memory.findMany({
+      where: { userId },
       orderBy: {
         createdAt: 'asc',
       },
@@ -18,20 +23,28 @@ export async function memoriesRoutes(app: FastifyInstance) {
       excerpt: memory.content.substring(0, 115).concat('...'),
     }))
   })
-  app.get('/memories/:id', async (request) => {
+  app.get('/memories/:id', async (request, replay) => {
+    const userId = request.user.sub;
+
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
 
     const { id } = paramsSchema.parse(request.params)
 
-    const memories = await prisma.memory.findUniqueOrThrow({
+    const memory = await prisma.memory.findUniqueOrThrow({
       where: { id },
     })
 
-    return memories
+    if(!memory.isPublic && memory.userId !== userId){
+      return replay.status(401).send();
+    }
+
+    return memory
   })
   app.post('/memories', async (request) => {
+    const userId = request.user.sub;
+
     const bodySchema = z.object({
       content: z.string(),
       coverUrl: z.string(),
@@ -55,7 +68,9 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     return memory
   })
-  app.put('/memories/:id', async (request) => {
+  app.put('/memories/:id', async (request, reply) => {
+    const userId = request.user.sub;
+
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -74,24 +89,41 @@ export async function memoriesRoutes(app: FastifyInstance) {
       isPublic
     } = bodySchema.parse(request.body);
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: { id }
+    })
+
+    if(memory.userId !== userId){
+      return reply.status(401).send();
+    }
+
+    memory = await prisma.memory.update({
       where: { id },
       data: {
         content,
         coverUrl,
-        isPublic,
-        userId
+        isPublic
       }
     })
 
     return memory
   })
-  app.delete('/memories/:id', async (request) => {
+  app.delete('/memories/:id', async (request, reply) => {
+    const userId = request.user.sub;
+
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
 
     const { id } = paramsSchema.parse(request.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: { id }
+    })
+
+    if(memory.userId !== userId){
+      return reply.status(401).send();
+    }
 
     await prisma.memory.delete({
       where: { id },
